backend/src/requirements.txt
Django==2.1.5
django-bootstrap-modal-forms==1.3.1
django-bootstrap4==0.0.7
django-formtools==2.1
pytz==2018.9
